+++ CHANGELOG.md
@@ -3,3 +3,5 @@ Changelog is normally kept [here](https://github.com/Dronehub/interceptor/releas
 This is only for new-version changes:
 
 # v2.2
+
+* added config backup and restore